Split timing into old and new code
